requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
httpx==0.28.1
python-multipart==0.0.6
alembic==1.14.0
stripe==11.4.1
slowapi==0.1.9

